Create number6.py
introduce function get_int() with parameter, called prompt, that will return an int and or continually ask the user for x until given a correct response.  The main function will define x and get it from the def get_int(prompt) function, , but will take your question in the main function and take that as prompt. Then main will ask for x, then hand it off to get_int(prompt) function to verify it is an int and then return it and then main will print the statement "x is whatever in x is. Using pass it will not show an error message but just continually prompt the user until getting the correct response.